Feature autodownload

+++ EnvyUpdate/DashboardPage.xaml.cs
@@ -210,6 +210,15 @@ namespace EnvyUpdate
 
                 if (skippedVer != onlineDriv)
                 {
+                    if (GlobalVars.autoDownload)
+                    {
+                        if (buttonDownload.IsVisible)
+                        {
+                            Debug.LogToFile("INFO Auto-Downloading driver.");
+                            buttonDownload_Click(null, null);
+                        }
+                    }
+
                     Debug.LogToFile("INFO Showing update popup notification.");
                     Notify.ShowDrivUpdatePopup();
                 }
@@ -340,25 +349,34 @@ namespace EnvyUpdate
 
         private void buttonDownload_Click(object sender, RoutedEventArgs e)
         {
-            progressbarDownload.Visibility = Visibility.Visible;
-            buttonDownload.IsEnabled = false;
-
-            if (File.Exists(Path.Combine(GlobalVars.saveDirectory, onlineDriv + "-nvidia-installer.exe.downloading")))
+            if (GlobalVars.isDownloading)
             {
-                Debug.LogToFile("WARN Found previous unfinished download, retrying.");
-                File.Delete(Path.Combine(GlobalVars.saveDirectory, onlineDriv + "-nvidia-installer.exe.downloading"));
+                Debug.LogToFile("WARN A download is already running.");
+                ShowSnackbar(Wpf.Ui.Common.ControlAppearance.Danger, Wpf.Ui.Common.SymbolRegular.ErrorCircle24, Properties.Resources.info_download_running, Properties.Resources.info_download_running_title);
             }
-            Thread thread = new Thread(() => {
-                using (WebClient client = new WebClient())
+            else
+            {
+                progressbarDownload.Visibility = Visibility.Visible;
+                buttonDownload.IsEnabled = false;
+                GlobalVars.isDownloading = true;
+
+                if (File.Exists(Path.Combine(GlobalVars.saveDirectory, onlineDriv + "-nvidia-installer.exe.downloading")))
                 {
-                    client.Headers["User-Agent"] = GlobalVars.useragent;
-                    client.DownloadProgressChanged += new DownloadProgressChangedEventHandler(client_DownloadProgressChanged);
-                    client.DownloadFileCompleted += new AsyncCompletedEventHandler(client_DownloadFileCompleted);
-                    client.DownloadFileAsync(new Uri(Util.GetDirectDownload(gpuURL)), Path.Combine(GlobalVars.saveDirectory, onlineDriv + "-nvidia-installer.exe.downloading"));
+                    Debug.LogToFile("WARN Found previous unfinished download, retrying.");
+                    File.Delete(Path.Combine(GlobalVars.saveDirectory, onlineDriv + "-nvidia-installer.exe.downloading"));
                 }
-            });
-            thread.Start();
-            Debug.LogToFile("INFO Started installer download.");
+                Thread thread = new Thread(() => {
+                    using (WebClient client = new WebClient())
+                    {
+                        client.Headers["User-Agent"] = GlobalVars.useragent;
+                        client.DownloadProgressChanged += new DownloadProgressChangedEventHandler(client_DownloadProgressChanged);
+                        client.DownloadFileCompleted += new AsyncCompletedEventHandler(client_DownloadFileCompleted);
+                        client.DownloadFileAsync(new Uri(Util.GetDirectDownload(gpuURL)), Path.Combine(GlobalVars.saveDirectory, onlineDriv + "-nvidia-installer.exe.downloading"));
+                    }
+                });
+                thread.Start();
+                Debug.LogToFile("INFO Started installer download.");
+            }
         }
 
         void client_DownloadProgressChanged(object sender, DownloadProgressChangedEventArgs e)
@@ -376,6 +394,7 @@ namespace EnvyUpdate
             Application.Current.Dispatcher.Invoke(new Action(() => {
                 buttonDownload.IsEnabled = true;
                 progressbarDownload.Visibility = Visibility.Collapsed;
+                GlobalVars.isDownloading = false;
             }));
             if (e.Error == null)
             {
